Actually sort the spreadsheet "date" column
A lot of internal logic depend on the correct sorting of this column
and so far it has not been done correctly.

Incorrect calculations were made if the spreadsheet "date" column was
not by default sorted in its original form.

Previously, as defined in the spreadsheet:

['2022-03-02 00:00:00', '2022-02-01 00:00:00', '2022-02-01 00:00:00',
 '2022-01-02 00:00:00', '2021-11-09 00:00:00', '2021-10-01 00:00:00',
                                [...]
 '2019-01-01 00:00:00', '2018-01-10 00:00:00', '2018-01-01 00:00:00']

Now with the same spreadsheet:

['2018-01-01 00:00:00', '2018-01-10 00:00:00', '2019-01-10 00:00:00',
                                [...]
 '2021-10-01 00:00:00', '2021-10-09 00:00:00', '2021-01-02 00:00:00',
 '2022-02-01 00:00:00', '2022-02-01 00:00:00', '2022-03-02 00:00:00']

Original idea came from ph03:
https://github.com/stadlmax/pyFIFOtax/pull/1/commits/3ea152bea88c8291304a589615641e8406e1f121