Rename "wire_transfers" sheet to "currency conversion to EUR"
"Wire transfers" were a misnomer. It assumed that the currency
conversion happened on the very same day the wire transfer was initiated
(or finished) which was rarely the case.

Also, it carried a hidden notation that the user must have converted the
USD to EUR after a wire transfer which wasn't always the case.

Hence, the naming changed to be a lot more clear on the indended purpose
of the sheet.

Backwards compatibility was kept by still enabling reading from the old
sheet name with a warning message