Pop the whole amount where it's close to the total
Due to floating point representation it's possible that the quantity to
be popped and the total quantity is the same but represented
differently.

In one case the total can be slightly less and triggers the first
condition (to_be_popped > total) which doesn't remove the amount from
the queue, leaving 0 quantities in the final report.

Swap the condition to check for close equality first.